+++ spring-integration-core/src/main/java/org/springframework/integration/aggregator/AbstractCorrelatingMessageHandler.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2014 the original author or authors.
+ * Copyright 2002-2015 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -461,14 +461,8 @@ public abstract class AbstractCorrelatingMessageHandler extends AbstractMessageP
 		if (this.discardChannelName != null) {
 			synchronized (this) {
 				if (this.discardChannelName != null) {
-					try {
-						this.discardChannel = getBeanFactory().getBean(this.discardChannelName, MessageChannel.class);
-						this.discardChannelName = null;
-					}
-					catch (BeansException e) {
-						throw new DestinationResolutionException("Failed to look up MessageChannel with name '"
-								+ this.discardChannelName + "' in the BeanFactory.");
-					}
+					this.discardChannel = getChannelResolver().resolveDestination(this.discardChannelName);
+					this.discardChannelName = null;
 				}
 			}
 		}
