+++ spring-integration-core/src/main/java/org/springframework/integration/router/AbstractMappingMessageRouter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2014 the original author or authors.
+ * Copyright 2002-2015 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -46,13 +46,12 @@ import org.springframework.util.StringUtils;
  * @author Oleg Zhurakousky
  * @author Gunnar Hillert
  * @author Gary Russell
+ * @author Artem Bilan
  * @since 2.1
  */
 public abstract class AbstractMappingMessageRouter extends AbstractMessageRouter implements MappingMessageRouterManagement {
 
 	private volatile Map<String, String> channelMappings = new ConcurrentHashMap<String, String>();
-
-	private volatile DestinationResolver<MessageChannel> channelResolver;
 
 	private volatile String prefix;
 
@@ -64,7 +63,6 @@ public abstract class AbstractMappingMessageRouter extends AbstractMessageRouter
 	/**
 	 * Provide mappings from channel keys to channel names.
 	 * Channel names will be resolved by the {@link DestinationResolver}.
-	 *
 	 * @param channelMappings The channel mappings.
 	 */
 	@Override
@@ -77,21 +75,7 @@ public abstract class AbstractMappingMessageRouter extends AbstractMessageRouter
 	}
 
 	/**
-	 * Specify the {@link DestinationResolver} strategy to use.
-	 * The default is a BeanFactoryChannelResolver.
-	 * This is considered an infrastructural configuration option and
-	 * as of 2.1 has been deprecated as a configuration-driven attribute.
-	 *
-	 * @param channelResolver The channel resolver.
-	 */
-	public void setChannelResolver(DestinationResolver<MessageChannel> channelResolver) {
-		Assert.notNull(channelResolver, "'channelResolver' must not be null");
-		this.channelResolver = channelResolver;
-	}
-
-	/**
 	 * Specify a prefix to be added to each channel name prior to resolution.
-	 *
 	 * @param prefix The prefix.
 	 */
 	public void setPrefix(String prefix) {
@@ -100,7 +84,6 @@ public abstract class AbstractMappingMessageRouter extends AbstractMessageRouter
 
 	/**
 	 * Specify a suffix to be added to each channel name prior to resolution.
-	 *
 	 * @param suffix The suffix.
 	 */
 	public void setSuffix(String suffix) {
@@ -110,7 +93,6 @@ public abstract class AbstractMappingMessageRouter extends AbstractMessageRouter
 	/**
 	 * Specify whether this router should ignore any failure to resolve a channel name to
 	 * an actual MessageChannel instance when delegating to the ChannelResolver strategy.
-	 *
 	 * @param resolutionRequired true if resolution is required.
 	 */
 	public void setResolutionRequired(boolean resolutionRequired) {
@@ -120,7 +102,6 @@ public abstract class AbstractMappingMessageRouter extends AbstractMessageRouter
 	/**
 	 * Returns an unmodifiable version of the channel mappings.
 	 * This is intended for use by subclasses only.
-	 *
 	 * @return The channel mappings.
 	 */
 	@Override
@@ -131,7 +112,6 @@ public abstract class AbstractMappingMessageRouter extends AbstractMessageRouter
 
 	/**
 	 * Add a channel mapping from the provided key to channel name.
-	 *
 	 * @param key The key.
 	 * @param channelName The channel name.
 	 */
@@ -143,7 +123,6 @@ public abstract class AbstractMappingMessageRouter extends AbstractMessageRouter
 
 	/**
 	 * Remove a channel mapping for the given key if present.
-	 *
 	 * @param key The key.
 	 */
 	@Override
@@ -152,25 +131,10 @@ public abstract class AbstractMappingMessageRouter extends AbstractMessageRouter
 		this.channelMappings.remove(key);
 	}
 
-	@Override
-	public void onInit() {
-		try {
-			super.onInit();
-		}
-		catch (Exception e) {
-			throw new IllegalStateException(e);
-		}
-		BeanFactory beanFactory = this.getBeanFactory();
-		if (this.channelResolver == null && beanFactory != null) {
-			this.channelResolver = new BeanFactoryChannelResolver(beanFactory);
-		}
-	}
-
 	/**
 	 * Subclasses must implement this method to return the channel keys.
 	 * A "key" might be present in this router's "channelMappings", or it
 	 * could be the channel's name or even the Message Channel instance itself.
-	 *
 	 * @param message The message.
 	 * @return The channel keys.
 	 */
@@ -193,7 +157,6 @@ public abstract class AbstractMappingMessageRouter extends AbstractMessageRouter
 	 * <p>Mappings must be delimited with newlines, for example:
 	 * <p>{@code "@'myRouter.handler'.replaceChannelMappings('foo=qux \n baz=bar')"}.
 	 * @param channelMappings The channel mappings.
-	 *
 	 * @since 4.0
 	 */
 	@Override
@@ -218,13 +181,9 @@ public abstract class AbstractMappingMessageRouter extends AbstractMessageRouter
 	}
 
 	private MessageChannel resolveChannelForName(String channelName, Message<?> message) {
-		if (this.channelResolver == null) {
-			this.onInit();
-		}
-		Assert.state(this.channelResolver != null, "unable to resolve channel names, no ChannelResolver available");
 		MessageChannel channel = null;
 		try {
-			channel = this.channelResolver.resolveDestination(channelName);
+			channel = getChannelResolver().resolveDestination(channelName);
 		}
 		catch (DestinationResolutionException e) {
 			if (this.resolutionRequired) {
