+++ spring-integration-core/src/main/java/org/springframework/integration/router/AbstractMessageProcessingRouter.java
@@ -46,7 +46,7 @@ class AbstractMessageProcessingRouter extends AbstractMappingMessageRouter
 
 
 	@Override
-	public final void onInit() {
+	public final void onInit() throws Exception {
 		super.onInit();
 		if (this.messageProcessor instanceof AbstractMessageProcessor) {
 			((AbstractMessageProcessor<?>) this.messageProcessor).setConversionService(this.getConversionService());
